Add uniprot isoform list for hg19
There were several transcripts that were not in the ensembl hg19 transcript
dump. This assigns an ensembl transcript for 281 of the 307 hugo symbols (the
other ones don't have a good alternative in hg19 dump). None of these 307 are
particularly import genes for our purposes. See analysis:

https://github.com/inodb/genome-nexus-importer/blob/one-geneid-per-hgnc/notebooks/ensembl_identifiers.ipynb

For each of these genes we pick in this order:

1. the one assigned by hgnc website
2. the one assigned with is_canonical is true in ensembl data dump
3. the longest protein
4. the latest gene_stable_id

See the google docs for all the assignments (ensembl_canonical_transcripts vs
uniprot_canonical_transcript):

https://docs.google.com/spreadsheets/d/12o3OhlT-K0XQHW6T_kRxCaWIwwUyYlUY5OZkQP9wwns/edit#gid=934483692

We also remove the 294 hugo symbols that weren't assignable using hg19 dump